Refactored classes and configurations and alos removd php-cs-fixer risky rules from the default rule sets.
